chore: rename createPR to githubPR
